Find Doctors and Patients

+++ src/main/java/com/jathiswar/hospital/dao/PatientPaymentDetailRepository.java
@@ -4,5 +4,5 @@ package com.jathiswar.hospital.dao;
 import com.jathiswar.hospital.entity.PatientPaymentDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientPaymentDetailRepository extends JpaRepository<PatientPaymentDetail,Integer> {
+public interface PatientPaymentDetailRepository extends JpaRepository<PatientPaymentDetail,Long> {
 }
